pyproject.toml
[project]
name = "compute-eval"
version = "0.1.0"
description = "Library for evaluating Large Language Models on CUDA code"
readme = "README.md"
authors = [
    { name = "NVIDIA" }
]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.42.0",
    "fire>=0.7.0",
    "numpy>=2.2.0",
    "openai>=1.58.1",
    "pre-commit>=4.0.1",
    "psutil>=6.1.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]

[tool.poetry]
name = "compute-eval"
version = "0.1.0"
description = "Library for evaluating Large Language Models on CUDA code"
authors = ["NVIDIA"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
compute_eval = "compute_eval.main:main"
